ws-srv/requirements.txt
python_socketio==5.1.0
uvicorn[standard]==0.13.1
websockets==8.1
Pika==0.13.0
